requirements.txt
pandas
Levenshtein
scikit-learn
markdown
markdown-it-py
bs4